requirements.txt
langgraph
langchain_openai
langchain_community
docx2txt
langchain_huggingface
python-dotenv
langchain-qdrant
guardrails-ai>=0.5.13
ragas